addressed bot comments
