+++ packages/visual-editor/src/vite-plugin/writeFileIfChanged.ts
@@ -1,0 +1,16 @@
+import fs from "fs-extra";
+
+export const writeFileIfChanged = (
+  filePath: string,
+  content: string
+): boolean => {
+  if (fs.existsSync(filePath)) {
+    const existingContent = fs.readFileSync(filePath, "utf8");
+    if (existingContent === content) {
+      return false;
+    }
+  }
+
+  fs.writeFileSync(filePath, content);
+  return true;
+};
